Refactor: move interfaces to the given directories
