+++ src/Drahak/Restful/Mapping/JsonMapper.php
@@ -1,7 +1,6 @@
 <?php
 namespace Drahak\Restful\Mapping;
 
-use Drahak\Restful\IMapper;
 use Drahak\Restful\InvalidArgumentException;
 use Nette\Object;
 use Nette\Utils\Json;
